+++ src/engine/validation_loop.py
@@ -12,7 +12,7 @@ def val_loop(model, loss_module, val_dataloader):
     model.eval()
     with torch.no_grad():
         for ventana, frame in val_dataloader:
-            ventana, frame = ventana.to(device), frame.to(device)
+            ventana, frame = ventana.to(device, dtype = torch.float32), frame.to(device, dtype = torch.float32)
             pred = model(ventana)
             loss =  loss_module(pred, frame, a = 5, b = 15)
             val_loss += loss.item()
